Completed Day 03 Part 2 challenge
Initially had an issue trying to use a large 2D array to keep track
of the grid state as I was processing the input wires. As I made
the grid larger, my program was running out of memory. I needed to
think of another way to keep track of the wire states without using
a large 2D array.

To avoid using a large 2D array, I used hash-maps to keep track of
each move made by the wire by mapping the simulated grid location
with the number of moves taken by the wire to get to that point. If
the wire overlaps itself, the first value isn't recorded so only
the minimum moves taken to get to the particular grid location is
recorded.

I also rewrote the Day 03 Part 1 solution to make use of this new
manner of processing the wires and keeping track of the problem
state.

Added some unit tests to the day_03 module to test my solutions
against the example input and answers given.