Completed Day 07 Part 1 challenge
Found the bug with my IntcodeMachine code. I was using a Vec<i32> to
store the input values for the machine, and using a .pop() operation to
get the next input value to be used when executing the program stored in
its memory. This resulted in the input values being used in the reverse
order to what was required.

To solve this, I decided to use a quere data structure instead for
storing the IntcodeMachine input values. This allowed for the oldest
value (i.e. the first in line) to be removed and used when called for
during program execution.