Completed Day 07 Part 2 challenge
Changed the data structure storing the output values in the
IntcodeMachine back to a queue. The .get_output() function performs
a peek at the head value of the output queue, without modifying the
values. Also added another function to clear the output queue, so
that modification of the output queue is not carried out in a
function that should be performing read-only functions.