Still working on Day 09 Part 1 - not yet finished.
Code passes the example tests given on the challenge page, but
gives output "203" when run against the puzzle input. Some research
on the challenge threads online indicates this "203" error might be
a common occurrence for people attempting this challenge. Will need
to look deeper at this to fix my code.