Completed Day 9 Part 1 challenge
Now this was an interesting one. After implementing the additional
functionality in IntcodeMachine required for the challenge, I
still had a bug, meaning I was getting the output "203" when
running my code on the puzzle input.

Did some looking around online and found some suggested additional
test cases on reddit - https://bit.ly/35pOx2s. Added these and ran
them, with all passing except last one. Had a closer look and this
indicated to me I was getting the output address during
OPCODE_INPUT incorrectly. I wasn't adding the relative base to the
output address when the parameter was in relative mode.

I fixed the issue by modifying the function used to retrieve a
parameter value based on its mode. The additional argument was
used to specify if a memory lookup should be performed or if the
calculated address value should be given. This was done to allow
the output address to be used to store the required value in
memory as needed by the particular opcode.

Made the change to all opcodes requiring a write to the memory
of the intcode machine.

Also decreased the amount of memory allocated for the intcode
machine to 10000 units (8 bytes per unit - storing one i64 value).