Completed Day 10 Part 2 challenge
This was an interesting one. Worked out I needed to reverse the order of the
y-axis when determining what quadrant an endpoint was in relative to start
location. This allowed me to calculate the asteroid angles for sorting. Also
needed to make an adjustment to the part 1 challenge code so the quandrant
multiple for y-axis as also inverted where needed.